hw1/requirements.txt
apache-airflow==2.10.5
apache-airflow-providers-sqlite==4.0.1
apache-airflow-providers-http==5.2.1
